+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Parameter `--docker-timeout` may also be set in the artifact
+  configuration file.
+
 ## [1.3.0] 2020-05-30
 
 ### Added